+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KFAQP8AWWF83VG4RGG7FP4PZ"
+        "01KFB0S0ZANBY50S6B3SV7WG07"
     )
     r"""The unique identifier of the data source"""