+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -148,7 +148,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KFAQP8ATE2PF1FS13BRRBXJJ"
+        "01KFB0S0Z735GV88RR0DWPQH5K"
     )
     r"""The unique identifier of the data source"""