+++ packages/orq-rc/src/orq_ai_sdk/models/listidentitiesop.py
@@ -208,7 +208,7 @@ class ListIdentitiesData(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-01-19T08:59:21.884Z")
+    updated: Optional[datetime] = parse_datetime("2026-01-19T11:38:09.062Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")