+++ packages/orq-rc/RELEASES.md
@@ -4769,3 +4769,13 @@ Based on:
 - [python v4.2.0-rc.39] packages/orq-rc
 ### Releases
 - [PyPI v4.2.0-rc.39] https://pypi.org/project/orq-ai-sdk/4.2.0-rc.39 - packages/orq-rc
+
+## 2026-01-19 11:40:49
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.688.0 (2.796.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.2.0-rc.40] packages/orq-rc
+### Releases
+- [PyPI v4.2.0-rc.40] https://pypi.org/project/orq-ai-sdk/4.2.0-rc.40 - packages/orq-rc